Split coordinator
